Created semester-based pages
I created one page for each semester (including the current semester), as well as a template page. index.html will continue to be the page for the current semester, whenever is current. Ideally, at the end of every semester, someone manually archives the content of index.html, makes all the relevant edits, and puts it in another file.